requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.2
django-filter==2.4.0
django-google-maps==0.12.4
django-guardian==2.3.0
django-otp==1.0.6
djangorestframework==3.12.2
googlemaps==4.4.5
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.0.0
pycodestyle==2.6.0
Pygments==2.8.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyotp==2.6.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
twilio==6.62.0
urllib3==1.26.4
wrapt==1.12.1